backend/requirements.txt
Django>=5.0
djangorestframework>=3.16.0
psycopg[binary]>=3.2.0
django-environ>=0.12.0
celery[redis]>=5.5.0
redis>=5.2.0
boto3>=1.40.0
django-storages>=1.14.0
qdrant-client>=1.15.0
pgvector>=0.4.0
numpy>=2.3.0
feedparser>=6.0.10
httpx>=0.25.0
beautifulsoup4>=4.12.0
langdetect>=1.0.9
trafilatura>=1.6.0
requests>=2.31.0
sentence-transformers>=5.1.0
numpy>=2.3.0
